Render settings improvements
