+++ Assets/Fantasy Skybox FREE/Materials/Sunny 01B.mat
@@ -87,7 +87,7 @@ Material:
       data:
         first:
           name: _Rotation
-        second: 220.375931
+        second: 2109.18555
     m_Colors:
       data:
         first:
@@ -96,4 +96,4 @@ Material:
       data:
         first:
           name: _Tint
-        second: {r: .493590474, g: .808217585, b: 1, a: 1}
+        second: {r: .989393592, g: .956982315, b: .542115867, a: 1}
