+++ Assets/Materials/Effects/GlobalGradientFog.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a69d5838115474e7d9cee39395fa852c
+timeCreated: 1432381103
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
