+++ Assets/ScriptableObjects/TimeLightingSettingsData/GardenDayLighting.asset
@@ -1,0 +1,60 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 01817ead0d84c442c929fc7269f632db, type: 3}
+  m_Name: GardenDayLighting
+  m_EditorClassIdentifier: 
+  timeSettings:
+    skyColor: {r: .544117689, g: .867951334, b: 1, a: 1}
+    lightColor: {r: .765720069, g: 1, b: .691176474, a: 1}
+    lightIntensity: 1
+    fogGradient:
+      key0:
+        serializedVersion: 2
+        rgba: 16777215
+      key1:
+        serializedVersion: 2
+        rgba: 1803091811
+      key2:
+        serializedVersion: 2
+        rgba: 2063597491
+      key3:
+        serializedVersion: 2
+        rgba: 2475276287
+      key4:
+        serializedVersion: 2
+        rgba: 2894706687
+      key5:
+        serializedVersion: 2
+        rgba: 2072623103
+      key6:
+        serializedVersion: 2
+        rgba: 2072623103
+      key7:
+        serializedVersion: 2
+        rgba: 2072623103
+      ctime0: 0
+      ctime1: 15035
+      ctime2: 44718
+      ctime3: 65535
+      ctime4: 65535
+      ctime5: 65535
+      ctime6: 65535
+      ctime7: 65535
+      atime0: 0
+      atime1: 15420
+      atime2: 37201
+      atime3: 64957
+      atime4: 64957
+      atime5: 65535
+      atime6: 65535
+      atime7: 65535
+      m_NumColorKeys: 3
+      m_NumAlphaKeys: 4
