+++ Assets/ScriptableObjects/TimeLightingSettingsData/GardenDayLighting.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3bc4af232d9344ffd9a93a92f3348e5a
+timeCreated: 1432587051
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
