+++ Assets/ScriptableObjects/TimeLightingSettingsData/GardenNightLighting.asset
@@ -1,0 +1,61 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 01817ead0d84c442c929fc7269f632db, type: 3}
+  m_Name: GardenNightLighting
+  m_EditorClassIdentifier: 
+  timeSettings:
+    skyColor: {r: .259791523, g: .160575286, b: .404411793, a: 1}
+    lightColor: {r: .779310346, g: .529411793, b: 1, a: 1}
+    lightIntensity: .5
+    fogDensity: .100000001
+    fogGradient:
+      key0:
+        serializedVersion: 2
+        rgba: 16729193
+      key1:
+        serializedVersion: 2
+        rgba: 2561812838
+      key2:
+        serializedVersion: 2
+        rgba: 4294919343
+      key3:
+        serializedVersion: 2
+        rgba: 4294919273
+      key4:
+        serializedVersion: 2
+        rgba: 4294919273
+      key5:
+        serializedVersion: 2
+        rgba: 4294919273
+      key6:
+        serializedVersion: 2
+        rgba: 4294919273
+      key7:
+        serializedVersion: 2
+        rgba: 4294919273
+      ctime0: 0
+      ctime1: 10409
+      ctime2: 26407
+      ctime3: 57343
+      ctime4: 57343
+      ctime5: 57343
+      ctime6: 57343
+      ctime7: 57343
+      atime0: 0
+      atime1: 6168
+      atime2: 53199
+      atime3: 57343
+      atime4: 57343
+      atime5: 57343
+      atime6: 57343
+      atime7: 57343
+      m_NumColorKeys: 3
+      m_NumAlphaKeys: 3
