+++ Assets/ScriptableObjects/TimeLightingSettingsData/GardenNightLighting.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 732012bdb46724f429ed5e60aeb54ba9
+timeCreated: 1432586805
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
