+++ Assets/ScriptableObjects/TimeLightingSettingsData/SpireBottomDayLighting.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f9acba14c8a6749308fa8fc86d40d5e7
+timeCreated: 1432586815
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
