+++ Assets/ScriptableObjects/TimeLightingSettingsData/SpireBottomNightLighting.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: be6142048a57c462bacd5918e91fd4b5
+timeCreated: 1432586825
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
