+++ Assets/ScriptableObjects/TimeLightingSettingsData/SpireMiddleDayLighting.asset
@@ -1,0 +1,61 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 01817ead0d84c442c929fc7269f632db, type: 3}
+  m_Name: SpireMiddleDayLighting
+  m_EditorClassIdentifier: 
+  timeSettings:
+    skyColor: {r: .545098066, g: .945098102, b: 1, a: 1}
+    lightColor: {r: .980392218, g: 1, b: .847058892, a: 1}
+    lightIntensity: .5
+    fogDensity: .0500000007
+    fogGradient:
+      key0:
+        serializedVersion: 2
+        rgba: 16747930
+      key1:
+        serializedVersion: 2
+        rgba: 4294919273
+      key2:
+        serializedVersion: 2
+        rgba: 4294967295
+      key3:
+        serializedVersion: 2
+        rgba: 4294967295
+      key4:
+        serializedVersion: 2
+        rgba: 4294967295
+      key5:
+        serializedVersion: 2
+        rgba: 4294967295
+      key6:
+        serializedVersion: 2
+        rgba: 4294967295
+      key7:
+        serializedVersion: 2
+        rgba: 4294967295
+      ctime0: 0
+      ctime1: 65535
+      ctime2: 16191
+      ctime3: 24672
+      ctime4: 32768
+      ctime5: 34117
+      ctime6: 49151
+      ctime7: 57054
+      atime0: 14071
+      atime1: 65535
+      atime2: 57343
+      atime3: 57343
+      atime4: 57343
+      atime5: 57343
+      atime6: 57343
+      atime7: 57343
+      m_NumColorKeys: 2
+      m_NumAlphaKeys: 2
