+++ Assets/ScriptableObjects/TimeLightingSettingsData/SpireMiddleDayLighting.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7d75a35b19663418eba1da5d7791d9a2
+timeCreated: 1432587931
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
