+++ Assets/ScriptableObjects/TimeLightingSettingsData/SpireMiddleNightLighting.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8e452f2c9dc2147849618ca0f86456f4
+timeCreated: 1432586831
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
