+++ Assets/ScriptableObjects/TimeLightingSettingsData/SpireTopDayLighting.asset
@@ -1,0 +1,60 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 01817ead0d84c442c929fc7269f632db, type: 3}
+  m_Name: SpireTopDayLighting
+  m_EditorClassIdentifier: 
+  timeSettings:
+    skyColor: {r: 1, g: .968559861, b: .544117689, a: 1}
+    lightColor: {r: .97870183, g: 1, b: .845588207, a: 1}
+    lightIntensity: 1
+    fogGradient:
+      key0:
+        serializedVersion: 2
+        rgba: 3558399
+      key1:
+        serializedVersion: 2
+        rgba: 2643054591
+      key2:
+        serializedVersion: 2
+        rgba: 3179919359
+      key3:
+        serializedVersion: 2
+        rgba: 3187659913
+      key4:
+        serializedVersion: 2
+        rgba: 2080363657
+      key5:
+        serializedVersion: 2
+        rgba: 2080363657
+      key6:
+        serializedVersion: 2
+        rgba: 3909080201
+      key7:
+        serializedVersion: 2
+        rgba: 3909080201
+      ctime0: 22359
+      ctime1: 46260
+      ctime2: 44911
+      ctime3: 57343
+      ctime4: 57343
+      ctime5: 57343
+      ctime6: 57343
+      ctime7: 57343
+      atime0: 0
+      atime1: 12143
+      atime2: 65535
+      atime3: 65535
+      atime4: 65535
+      atime5: 65535
+      atime6: 57343
+      atime7: 57343
+      m_NumColorKeys: 2
+      m_NumAlphaKeys: 3
