+++ Assets/ScriptableObjects/TimeLightingSettingsData/SpireTopDayLighting.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d8230e05c38c847d8b009b450fc5f4ee
+timeCreated: 1432586848
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
