+++ Assets/ScriptableObjects/TimeLightingSettingsData/SpireTopNightLighting.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a918d534c613446e0899aa7ca9a51a48
+timeCreated: 1432586856
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
