+++ Assets/ScriptableObjects/TimeLightingSettingsData/TutorialDayLighting.asset
@@ -11,50 +11,50 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 01817ead0d84c442c929fc7269f632db, type: 3}
   m_Name: TutorialDayLighting
   m_EditorClassIdentifier: 
-  skyColor: {r: .492647052, g: .811054587, b: 1, a: 1}
-  lightColor: {r: 1, g: .939655185, b: .742647052, a: 1}
-  lightIntensity: 1
-  fogDensity: .00999999978
-  fogGradient:
-    key0:
-      serializedVersion: 2
-      rgba: 16777215
-    key1:
-      serializedVersion: 2
-      rgba: 577883244
-    key2:
-      serializedVersion: 2
-      rgba: 2218427186
-    key3:
-      serializedVersion: 2
-      rgba: 2411237888
-    key4:
-      serializedVersion: 2
-      rgba: 4290286080
-    key5:
-      serializedVersion: 2
-      rgba: 553648127
-    key6:
-      serializedVersion: 2
-      rgba: 553648127
-    key7:
-      serializedVersion: 2
-      rgba: 553648127
-    ctime0: 0
-    ctime1: 14071
-    ctime2: 31804
-    ctime3: 61680
-    ctime4: 65535
-    ctime5: 40959
-    ctime6: 49151
-    ctime7: 57632
-    atime0: 0
-    atime1: 11372
-    atime2: 22166
-    atime3: 47417
-    atime4: 65535
-    atime5: 57343
-    atime6: 57343
-    atime7: 57343
-    m_NumColorKeys: 5
-    m_NumAlphaKeys: 5
+  timeSettings:
+    skyColor: {r: 1, g: .742190719, b: .544117689, a: 1}
+    lightColor: {r: .97870183, g: 1, b: .845588207, a: 1}
+    lightIntensity: 1
+    fogGradient:
+      key0:
+        serializedVersion: 2
+        rgba: 16770714
+      key1:
+        serializedVersion: 2
+        rgba: 2122954751
+      key2:
+        serializedVersion: 2
+        rgba: 2080363657
+      key3:
+        serializedVersion: 2
+        rgba: 3909080201
+      key4:
+        serializedVersion: 2
+        rgba: 3909080201
+      key5:
+        serializedVersion: 2
+        rgba: 3909080201
+      key6:
+        serializedVersion: 2
+        rgba: 3909080201
+      key7:
+        serializedVersion: 2
+        rgba: 3909080201
+      ctime0: 0
+      ctime1: 22937
+      ctime2: 57343
+      ctime3: 57343
+      ctime4: 57343
+      ctime5: 57343
+      ctime6: 57343
+      ctime7: 57343
+      atime0: 0
+      atime1: 38550
+      atime2: 65535
+      atime3: 57343
+      atime4: 57343
+      atime5: 57343
+      atime6: 57343
+      atime7: 57343
+      m_NumColorKeys: 2
+      m_NumAlphaKeys: 3
