+++ Assets/ScriptableObjects/TimeLightingSettingsData/TutorialNightLighting.asset
@@ -11,50 +11,50 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 01817ead0d84c442c929fc7269f632db, type: 3}
   m_Name: TutorialNightLighting
   m_EditorClassIdentifier: 
-  skyColor: {r: .654766977, g: .323529422, b: 1, a: 1}
-  lightColor: {r: .768965542, g: .507352948, b: 1, a: 1}
-  lightIntensity: .300000012
-  fogDensity: .200000003
-  fogGradient:
-    key0:
-      serializedVersion: 2
-      rgba: 3726769920
-    key1:
-      serializedVersion: 2
-      rgba: 16777215
-    key2:
-      serializedVersion: 2
-      rgba: 486539263
-    key3:
-      serializedVersion: 2
-      rgba: 486539263
-    key4:
-      serializedVersion: 2
-      rgba: 486539263
-    key5:
-      serializedVersion: 2
-      rgba: 486539263
-    key6:
-      serializedVersion: 2
-      rgba: 486539263
-    key7:
-      serializedVersion: 2
-      rgba: 486539263
-    ctime0: 0
-    ctime1: 57343
-    ctime2: 57343
-    ctime3: 57343
-    ctime4: 57343
-    ctime5: 57343
-    ctime6: 57343
-    ctime7: 57343
-    atime0: 0
-    atime1: 57343
-    atime2: 57343
-    atime3: 57343
-    atime4: 57343
-    atime5: 57343
-    atime6: 57343
-    atime7: 57343
-    m_NumColorKeys: 2
-    m_NumAlphaKeys: 2
+  timeSettings:
+    skyColor: {r: .404411793, g: .176940411, b: .124891877, a: 1}
+    lightColor: {r: 1, g: .831237316, b: .529411793, a: 1}
+    lightIntensity: .5
+    fogGradient:
+      key0:
+        serializedVersion: 2
+        rgba: 16777215
+      key1:
+        serializedVersion: 2
+        rgba: 3123273340
+      key2:
+        serializedVersion: 2
+        rgba: 3308664470
+      key3:
+        serializedVersion: 2
+        rgba: 4282664191
+      key4:
+        serializedVersion: 2
+        rgba: 4294919273
+      key5:
+        serializedVersion: 2
+        rgba: 4294919273
+      key6:
+        serializedVersion: 2
+        rgba: 4294919273
+      key7:
+        serializedVersion: 2
+        rgba: 4294919273
+      ctime0: 0
+      ctime1: 14071
+      ctime2: 41827
+      ctime3: 65535
+      ctime4: 57343
+      ctime5: 57343
+      ctime6: 57343
+      ctime7: 57343
+      atime0: 0
+      atime1: 33153
+      atime2: 65535
+      atime3: 57343
+      atime4: 57343
+      atime5: 57343
+      atime6: 57343
+      atime7: 57343
+      m_NumColorKeys: 3
+      m_NumAlphaKeys: 3
