+++ Assets/Scripts/Moving Objects/Ribbon.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 43c1face099274dc78337330676bc21a
+folderAsset: yes
+timeCreated: 1430472595
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
