+++ Assets/Scripts/VFX/RenderSettings/TimeLightingSettings.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 510bfc8f341584102a63b8eb9f4fb359
-timeCreated: 1432380091
+guid: a01eb78693c7e460db7ade0d2f96de1b
+timeCreated: 1432533795
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
