--- Assets/Shaders/VFX/GlobalFog.shader.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 9b11c4a942b4c41039461031e11fd7a1
-timeCreated: 1428211753
-licenseType: Free
-ShaderImporter:
-  defaultTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
