--- Assets/Standard Assets/Effects/ImageEffects/Scripts/GlobalFog.cs.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 76b5ec6153a1d55438228df10fe66844
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences:
-  - fogShader: {fileID: 4800000, guid: 70d8568987ac0499f952b54c7c13e265, type: 3}
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName:
